+++ order/Order.schema.ts
@@ -1,5 +1,4 @@
 import {Schema, model, Types} from 'mongoose'
-import {Product} from "./Product"
 
 const schema = new Schema({
     user: {type: Types.ObjectId, ref: 'User'},
