+++ order/order.validators.ts
@@ -1,19 +1,6 @@
 import {body} from "express-validator"
-import {Product} from "../models/Product"
-import {User} from "../models/User"
-
-export const userValidators = [
-    body('name').isLength({min: 4}),
-    body('phone').isMobilePhone("uk-UA", undefined),
-    body('email').isEmail(),
-    body('password').isLength({min: 6})
-]
-
-export const productValidators = [
-    body('name').isLength({min: 4}),
-    body('price').isFloat({min: 0}),
-    body('isAvailable').isBoolean()
-]
+import {Product} from "../product/Product.schema"
+import {User} from "../user/User.schema"
 
 export const orderValidators = [
     body('user')
